Store onboarded state in redux
