+++ package.json
@@ -33,6 +33,7 @@
     "expo-crypto": "~12.4.1",
     "expo-dev-client": "~2.4.6",
     "expo-device": "~5.4.0",
+    "expo-haptics": "~12.4.0",
     "expo-image": "~1.3.2",
     "expo-screen-orientation": "~6.0.5",
     "expo-sensors": "~12.3.0",
@@ -65,8 +66,8 @@
     "react-navigation-stack": "^2.10.4",
     "react-redux": "^8.0.2",
     "redux-persist": "^6.0.0",
-    "typescript": "^5.1.6",
-    "expo-haptics": "~12.4.0"
+    "semver": "^7.5.4",
+    "typescript": "^5.1.6"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
