+++ redux/SettingsSlice.ts
@@ -1,15 +1,19 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { SemVer, valid } from 'semver';
+import * as Application from 'expo-application';
 
 interface SettingsState {
     home_fullscreen: boolean;
     multiplier: number;
     currentGameId: string | undefined;
-}
+    onboarded: string | undefined;
+};
 
 const initialState: SettingsState = {
     home_fullscreen: false,
     multiplier: 1,
     currentGameId: undefined,
+    onboarded: undefined,
 };
 
 const settingsSlice = createSlice({
@@ -25,6 +29,11 @@ const settingsSlice = createSlice({
         setMultiplier(state, action: PayloadAction<number>) {
             state.multiplier = action.payload;
         },
+        setOnboardedVersion(state) {
+            const appVersion = new SemVer(Application.nativeApplicationVersion || '0.0.0');
+            console.log(`Setting Onboarded Version: ${appVersion}`);
+            state.onboarded = valid(appVersion) || '0.0.0';
+        }
     }
 });
 
@@ -32,6 +41,7 @@ export const {
     setCurrentGameId,
     toggleHomeFullscreen,
     setMultiplier,
+    setOnboardedVersion,
 } = settingsSlice.actions;
 
 export default settingsSlice.reducer;
